Add TFDV worker harness image (with build) (#90)
TFDV ships with native code which the Dataflow/Beam workers require access to. Therefore, we need a custom worker harness image with that native code.

This change also adds a CircleCI build step which pushes the harness image to a Docker repo on GCR. Once the first image is pushed, I'll make it public. It hasn't been published yet because CircleCI secrets are not available on non-master branches.