luigi.tfdv: Stats are placed in same dir as data
We have a number of examples internally where the output target which we
would like to run TFDV over, actually has multiple directories of data
inside. For example (to simplify), within the same GCSTarget there may
be a directory called "training" and "evaluation".

In the following example, the `_stats.pb` will now be placed in the same
directory as the `part-` files.

```
class UpstreamJobStats(TFDVGenerateStatsTask):
    def requires(self):
        return UpstreamJob()

    def file_pattern(self):
        return {'input': 'some/dir/part-*'}
```